Make project file classifier generally available
Previously this classification was localised to the import tree.

This commit moves the class to an outer scope and gives a more generally descriptive name.

We want to use this type to help classify files as part of an up-to-date check optmisation in a subsequent commit.
